requirement.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
flask>=2.0.1
Werkzeug>=2.0.1
paddleocr
paddlepaddle
google-generativeai>=0.3.0
azure-cognitiveservices-vision-computervision>=0.9.0
azure-ai-translation-text>=1.0.0
azure-core>=1.26.0
msrest>=0.7.1
python-dotenv>=0.19.0